Add some explicit cast to fix warning on 32 bit archs.
Yeah we could use %z here nowadays but we don't have strict C99 requirements(?),
well we have some :D

[skip ci]

+++ src/url_unreal.c
@@ -560,7 +560,7 @@ int https_connect_send_header(Download *handle)
 					    VERSIONONLY,
 					    hostandport,
 					    add_default_content_type ? "Content-Type: application/x-www-form-urlencoded\r\n" : "",
-					    strlen(handle->body));
+					    (long)strlen(handle->body));
 		}
 	} else
 		abort();
